refactor(billing): CHECKOUT-2951 Convert `billing` module into TypeScript
